requirements.txt
asgiref==3.8.1
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1.2
django-autoslug==1.9.9
django-countries==7.6.1
django-environ==0.11.2
django-filter==24.3
django-phonenumber-field==8.0.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
flake8==7.1.1
idna==3.10
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.47
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
factory_boy==3.3.1
Faker==30.6.0
inflection==0.5.1
iniconfig==2.0.0
pluggy==1.5.0
pytest==8.3.3
pytest-django==4.9.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
six==1.16.0
pytest-cov==5.0.0
celery==5.4.0
redis==5.1.1
flower==2.0.1
django-celery-email==3.0.0
argh==0.31.3
PyYAML==6.0.2
watchdog==5.0.3